Fix build fail: ambiguous fprintf() arguments Fixes bug #452498 (Theppitak
2007-07-05  Olav Vitters  <olav@bkor.dhs.org>

* common/gkr-secure-memory.c: (suba_print_cell):
* library/gnome-keyring.c: (read_all): Fix build fail: ambiguous
fprintf() arguments Fixes bug #452498 (Theppitak Karoonboonyanan).

svn path=/trunk/; revision=676

+++ common/gkr-secure-memory.c
@@ -340,9 +340,13 @@ suba_print_cell(struct allocator *suba, const char *msg, struct cell *c)
 {
 	ref_t ref = suba_ref(suba, c);
 	if (ref >= ALIGN(sizeof *suba) && (ref + POFF + c->size) <= 10000000) {
-		fprintf(stderr, "%s: %8u-%-8u %8u %-8u\n", msg, ref, ref + POFF + c->size, c->size, c->next);
+		fprintf(stderr, "%s: %8u-%-8u %8u %-8u\n", msg,
+			(unsigned int)ref, (unsigned int)(ref + POFF + c->size),
+			(unsigned int)c->size, (unsigned int)c->next);
 	} else {
-		fprintf(stderr, "%s: %8u-err %8u %-8u\n", msg, ref, c->size, c->next);
+		fprintf(stderr, "%s: %8u-err %8u %-8u\n", msg,
+			(unsigned int)ref, (unsigned int)c->size,
+			(unsigned int)c->next);
 		return 0;
 	}
 	return 1;
